+++ tests/api/documents/test_documents_add_cover_url_extension.py
@@ -22,7 +22,7 @@ from unittest import mock
 from rero_ils.modules.documents.extensions.add_cover_url import AddCoverUrlExtension
 
 
-def test_add_cover_url_extension_with_isbn(app, thumbnail_covers, cover_url):
+def test_add_cover_url_extension_with_isbn(appctx, thumbnail_covers, cover_url):
     """Test extension adds cover URL when ISBN is present."""
     thumbnail_covers["9781234567890"] = cover_url
     extension = AddCoverUrlExtension(cached=False)
@@ -36,10 +36,10 @@ def test_add_cover_url_extension_with_isbn(app, thumbnail_covers, cover_url):
     assert locator["type"] == "relatedResource"
     assert locator["content"] == "coverImage"
     assert locator["url"] == cover_url
-    assert locator["note"] == "rero-invenio-thumbnails provider: test"
+    assert "rero-invenio-thumbnails provider: test" in locator["publicNote"]
 
 
-def test_add_cover_url_extension_without_isbn(app, thumbnail_covers):
+def test_add_cover_url_extension_without_isbn(appctx, thumbnail_covers):
     """Test extension does nothing when no ISBN is present."""
     extension = AddCoverUrlExtension(cached=False)
     document = {"pid": "doc1", "identifiedBy": [{"type": "bf:Issn", "value": "1234-5678"}]}
@@ -50,7 +50,7 @@ def test_add_cover_url_extension_without_isbn(app, thumbnail_covers):
     assert "electronicLocator" not in document
 
 
-def test_add_cover_url_extension_already_has_cover(app, thumbnail_covers, cover_url):
+def test_add_cover_url_extension_already_has_cover(appctx, thumbnail_covers, cover_url):
     """Test extension does nothing when a coverImage locator already exists."""
     thumbnail_covers["9781234567890"] = cover_url
     extension = AddCoverUrlExtension(cached=False)
@@ -69,7 +69,7 @@ def test_add_cover_url_extension_already_has_cover(app, thumbnail_covers, cover_
     assert document["electronicLocator"][0]["url"] == "https://existing.example.com/cover.jpg"
 
 
-def test_add_cover_url_extension_multiple_isbns(app, thumbnail_covers, cover_url):
+def test_add_cover_url_extension_multiple_isbns(appctx, thumbnail_covers, cover_url):
     """Test extension uses first successful ISBN in sorted order."""
     # Only the second sorted ISBN ("9781234567890") has a cover
     thumbnail_covers["9781234567890"] = cover_url
@@ -90,9 +90,8 @@ def test_add_cover_url_extension_multiple_isbns(app, thumbnail_covers, cover_url
     assert document["electronicLocator"][0]["url"] == cover_url
 
 
-def test_add_cover_url_extension_no_thumbnail_available(app, thumbnail_covers):
+def test_add_cover_url_extension_no_thumbnail_available(appctx, thumbnail_covers):
     """Test extension does nothing when no thumbnail is available for any ISBN."""
-    # covers is empty — TestThumbnailProvider returns (None, "test") for all ISBNs
     extension = AddCoverUrlExtension(cached=False)
     document = {"pid": "doc1", "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}]}
 
@@ -102,19 +101,7 @@ def test_add_cover_url_extension_no_thumbnail_available(app, thumbnail_covers):
     assert "electronicLocator" not in document
 
 
-def test_add_cover_url_extension_cached_parameter(app, thumbnail_covers, cover_url):
-    """Test extension works correctly with cached=False."""
-    thumbnail_covers["9781234567890"] = cover_url
-    extension = AddCoverUrlExtension(cached=False)
-    document = {"pid": "doc1", "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}]}
-
-    result = extension.add_cover_url(document)
-
-    assert result is True
-    assert document["electronicLocator"][0]["url"] == cover_url
-
-
-def test_add_cover_url_extension_preserves_existing_locators(app, thumbnail_covers, cover_url):
+def test_add_cover_url_extension_preserves_existing_locators(appctx, thumbnail_covers, cover_url):
     """Test extension preserves existing non-cover electronicLocators."""
     thumbnail_covers["9781234567890"] = cover_url
     extension = AddCoverUrlExtension(cached=False)
@@ -131,9 +118,8 @@ def test_add_cover_url_extension_preserves_existing_locators(app, thumbnail_cove
     assert document["electronicLocator"][1]["content"] == "coverImage"
 
 
-def test_add_cover_url_extension_sorted_isbns(app, thumbnail_covers, cover_url):
+def test_add_cover_url_extension_sorted_isbns(appctx, thumbnail_covers, cover_url):
     """Test extension uses the first (smallest) ISBN that has a cover."""
-    # Only the smallest sorted ISBN has a cover — verifies sorted-order processing
     thumbnail_covers["9780000000000"] = cover_url
     extension = AddCoverUrlExtension(cached=False)
     document = {
@@ -150,7 +136,7 @@ def test_add_cover_url_extension_sorted_isbns(app, thumbnail_covers, cover_url):
     assert document["electronicLocator"][0]["url"] == cover_url
 
 
-def test_add_cover_url_extension_exception_then_success(app, cover_url):
+def test_add_cover_url_extension_exception_then_success(appctx, cover_url):
     """Test extension continues to next ISBN after get_thumbnail_url raises."""
     extension = AddCoverUrlExtension(cached=False)
     document = {
@@ -172,3 +158,80 @@ def test_add_cover_url_extension_exception_then_success(app, cover_url):
     assert result is True
     assert document["electronicLocator"][0]["url"] == cover_url
     assert mock_get.call_count == 2
+
+
+# --- post_create / post_commit hook tests ---
+
+
+def test_post_commit_enqueues_when_no_cover(appctx):
+    """post_commit enqueues the PID when the record has an ISBN and no cover."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {"pid": "doc1", "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}]}
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_commit(record)
+
+    mock_enqueue.assert_called_once_with("doc1", cached=False)
+
+
+def test_post_commit_skips_when_has_cover(appctx, cover_url):
+    """post_commit does not enqueue when the record already has a cover."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {
+        "pid": "doc1",
+        "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}],
+        "electronicLocator": [{"type": "relatedResource", "content": "coverImage", "url": cover_url}],
+    }
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_commit(record)
+
+    mock_enqueue.assert_not_called()
+
+
+def test_post_commit_skips_when_no_isbn(appctx):
+    """post_commit does not enqueue when the record has no ISBN."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {"pid": "doc1", "identifiedBy": [{"type": "bf:Issn", "value": "1234-5678"}]}
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_commit(record)
+
+    mock_enqueue.assert_not_called()
+
+
+def test_post_create_enqueues_when_no_cover(appctx):
+    """post_create enqueues the PID when the record has an ISBN and no cover."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {"pid": "doc1", "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}]}
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_create(record)
+
+    mock_enqueue.assert_called_once_with("doc1", cached=False)
+
+
+def test_post_create_skips_when_has_cover(appctx, cover_url):
+    """post_create does not enqueue when the record already has a cover."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {
+        "pid": "doc1",
+        "identifiedBy": [{"type": "bf:Isbn", "value": "9781234567890"}],
+        "electronicLocator": [{"type": "relatedResource", "content": "coverImage", "url": cover_url}],
+    }
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_create(record)
+
+    mock_enqueue.assert_not_called()
+
+
+def test_post_create_skips_when_no_isbn(appctx):
+    """post_create does not enqueue when the record has no ISBN."""
+    extension = AddCoverUrlExtension(cached=False)
+    record = {"pid": "doc1", "identifiedBy": [{"type": "bf:Issn", "value": "1234-5678"}]}
+
+    with mock.patch("rero_ils.modules.documents.extensions.add_cover_url.enqueue_cover_url_update") as mock_enqueue:
+        extension.post_create(record)
+
+    mock_enqueue.assert_not_called()
